// src/Models/Bureau.ts
export interface Bureau {
    id: number;
    bureauRaisonSocial: string;
    bureauRaisonSocialArabe: string;
    bureauPersonneContact: string;
    bureauPersonneContactArabe: string;
    bureauTelephone: string;
    bureauGsm: string;
    bureauFax: string;
    bureauAdresse: string;
    bureauAdresseArabe: string;
    bureauCodePostal: string;
    bureauEmail: string;
  }